Fix files for -Werror

+++ src/lib-tempo/src/component/ComponentAI.cpp
@@ -24,12 +24,11 @@ ComponentAI::ComponentAI(sf::Packet p)
 	uint32_t tmp;
 	uint32_t path_size;
 	p >> tmp;
-	type = (MoveType) tmp;
+	type = (MoveType)tmp;
 	p >> hitTeleport;
 	p >> hitSlide;
 	p >> path_size;
-	for (int I = 0; I < path_size; I++)
-	{
+	for (int I = 0; I < static_cast<int>(path_size); I++) {
 		glm::ivec2 tmp;
 		p >> tmp;
 		path.push_back(tmp);
@@ -45,17 +44,15 @@ sf::Packet ComponentAI::dumpComponent()
 {
 	sf::Packet p;
 
-	p << (uint32_t) type;
+	p << (uint32_t)type;
 	p << hitTeleport;
 	p << hitSlide;
-	p << (uint32_t) path.size();
-	for (int I = 0; I < path.size(); I++)
-	{
+	p << (uint32_t)path.size();
+	for (int I = 0; I < static_cast<int>(path.size()); I++) {
 		glm::ivec2 vec = path.front();
 		path.pop_front();
 		p << vec;
 		path.push_back(vec);
-
 	}
 	return p;  // does nothing
 }
