+++ src/lib-tempo/src/mask.cpp
@@ -2,10 +2,10 @@
 
 namespace tempo
 {
-Mask::Mask(glm::ivec2 centre, float *data, glm::ivec2 size)
+Mask::Mask(glm::ivec2 centre, float* data, glm::ivec2 size)
 {
 	ctr = centre;
-	sz  = size;
+	sz = size;
 
 	for (int I = 0; I < sz.x * sz.y; I++) {
 		mask.push_back(data[I]);
@@ -16,8 +16,8 @@ Mask::Mask(glm::ivec2 centre, float *data, glm::ivec2 size)
 
 Mask::Mask()
 {
-	ctr  = glm::ivec2(0, 0);
-	sz   = glm::ivec2(0, 0);
+	ctr = glm::ivec2(0, 0);
+	sz = glm::ivec2(0, 0);
 	mask = std::vector<float>();
 }
 
@@ -38,7 +38,7 @@ float Mask::get_value(int x, int y)
 	}
 }
 
-sf::Packet &operator<<(sf::Packet &packet, const tempo::Mask &m)
+sf::Packet& operator<<(sf::Packet& packet, const tempo::Mask& m)
 {
 	packet << m.ctr;
 	packet << m.sz;
@@ -50,14 +50,14 @@ sf::Packet &operator<<(sf::Packet &packet, const tempo::Mask &m)
 	uint32_t size = m.positions.size();
 	packet << size;
 
-	for (int I = 0; I < size; I++) {
+	for (int I = 0; I < static_cast<int>(size); I++) {
 		packet << m.positions[I];
 	}
 
 	return packet;
 }
 
-sf::Packet &operator>>(sf::Packet &packet, tempo::Mask &m)
+sf::Packet& operator>>(sf::Packet& packet, tempo::Mask& m)
 {
 	packet >> m.ctr;
 	packet >> m.sz;
@@ -75,7 +75,7 @@ sf::Packet &operator>>(sf::Packet &packet, tempo::Mask &m)
 
 	m.positions = std::vector<glm::ivec2>();
 
-	for (int I = 0; I < size; I++) {
+	for (int I = 0; I < static_cast<int>(size); I++) {
 		glm::ivec2 tmp_vec;
 		packet >> tmp_vec;
 		m.positions.push_back(tmp_vec);
