+++ src/lib-tempo/src/system/SystemTrigger.cpp
@@ -10,14 +10,14 @@
 namespace tempo
 {
 
-SystemTrigger::SystemTrigger(anax::World &world)
+SystemTrigger::SystemTrigger(anax::World& world)
 {
 	subSystem = SubSystemGetPlayers();
 	world.addSystem(subSystem);
 	world.refresh();
 }
 
-void SystemTrigger::updateButtons(anax::World &world)
+void SystemTrigger::updateButtons(anax::World& world)
 {
 	// find all player locations in game
 	playerPos = subSystem.getPlayers();
@@ -25,19 +25,19 @@ void SystemTrigger::updateButtons(anax::World &world)
 	// get all the button groups
 	auto entities = getEntities();
 
-	for (auto &entity : entities) {
+	for (auto& entity : entities) {
 		// get deque of all buttons in the group
-		auto &button_group = entity.getComponent<tempo::ComponentButtonGroup>();
-		std::deque<button> &buttons = button_group.buttons;
+		auto& button_group = entity.getComponent<tempo::ComponentButtonGroup>();
+		std::deque<button>& buttons = button_group.buttons;
 
 		// search through each button in a group to see if a player location matches
 		bool isGroupTriggered = true;
-		for (int i = 0; i < buttons.size(); i++) {
+		for (int i = 0; i < static_cast<float>(buttons.size()); i++) {
 			glm::ivec2 tempButtonPos = buttons[i].pos;
 
 			bool is_in = false;
 
-			for (int j = 0; j < playerPos.size(); j++) {
+			for (int j = 0; j < static_cast<int>(playerPos.size()); j++) {
 				if ((tempButtonPos.x == playerPos[j].x) && (tempButtonPos.y == playerPos[j].y)) {
 					is_in = true;
 					break;
@@ -52,12 +52,12 @@ void SystemTrigger::updateButtons(anax::World &world)
 			}
 		}
 		button_group.groupTriggered |= isGroupTriggered;
-		
+
 		if (button_group.groupTriggered == true && !button_group.action_happened) {
 			button_group.action_happened = true;
-			for (auto &entity : world.getEntities()) {
+			for (auto& entity : world.getEntities()) {
 				if (entity.hasComponent<tempo::ComponentStage>()) {
-					auto &component_stage = entity.getComponent<tempo::ComponentStage>();
+					auto& component_stage = entity.getComponent<tempo::ComponentStage>();
 					component_stage.setHeight(button_group.wall_positions, 0.f);
 				}
 			}
@@ -72,9 +72,9 @@ std::vector<glm::ivec2> SubSystemGetPlayers::getPlayers()
 	auto entities = getEntities();
 
 	// Get all the players and save their locations
-	for (auto &entity : entities) {
-		if (entity.hasComponent<tempo::ComponentPlayerLocal>()
-		    || entity.hasComponent<tempo::ComponentPlayerRemote>()) {
+	for (auto& entity : entities) {
+		if (entity.hasComponent<tempo::ComponentPlayerLocal>() ||
+		    entity.hasComponent<tempo::ComponentPlayerRemote>()) {
 			glm::ivec2 tempPlayerPos =
 			  entity.getComponent<tempo::ComponentStagePosition>().getOrigin();
 			currentPlayerPos.push_back(tempPlayerPos);
